requirements.txt
pandas
numpy
requests
python-dotenv
matplotlib
networkx
pytest
requests-mock
